Fixes the issue causing reboot instead of poweroff
The initramfs shutdown script now calls reboot and poweroff with the -f
flag. This fixes the issue with the system rebooting when the user
requested a poweroff. The issue appears to be that the sysvinit
reboot/poweroff can't detect the system runlevel (due to a lack of utmp)
and therefore doesn't know that it needs to actually perform the
reboot/poweroff instead of calling shutdown