Now runs fsck on source media at boot
If BOOTFSCK is enabled in config, the initramfs will try to run fsck on
the source media before mounting it. This copies some fsck tools to the
initramfs if enabled.

Right now, FAT and EXT? filesystems are supported, as those are the most
likely source filesystems.

NTFS is not currently supported, as there isn't really an NTFS-checking
tool for Linux (at least, not more than a fairly minimal one).

If the source filesystem doesn't have an fsck tool installed, the check
is bypassed.

If persistent changes are enabled, the persistence image also has fsck
run on it.