Initial implementation of persistent change snapshots
The boot option 'snapshot' may now be specified to take a snapshot of
the current persistent changes.

This is currently only tested with overlayfs. aufs is supported but
currently untested. Native changes are currently not supported.

This is done by mounting the persistent changes into a temporary
directory, squashing it into a bundle in the bundles directory (with an
incrementing prefix number in the 900s), then deleting the persistent
changes to be recreated.

This can be useful, for example, if the existing changes file is getting
full, or if the user wishes to test something on their system while
still being able to roll back changes.

These snapshots can be rolled back, by first deleting the persistent
changes image, and then deleting the snapshot bundles in reverse order,
until the desired state is restored.