snapshots no longer save /home or /tmp
Snapshots no longer save home folders, which now remain in the changes
image, nor /tmp, which is simply deleted.

This seems like a more useful mode of operation, as snapshots are really
intended to be an option to save system configuration, which is
relatively stable. User data, on the other hand, is constantly in flux,
and is probably best not snapshotted, as it would freeze a lot of
transient data in place where the space can never be (easily) reclaimed.

/tmp, obviously, is by definition transient garbage, and should just be
dumped.