Tweak kernel version detect code
The code blurb in config that detects the kernel in the target system
now redirects the stderr from the call to file(1) to /dev/null. This is
because the config file is loaded in the initramfs, which doesn't *have*
file, and thus generates an error. This suppresses it.

This also means the $KERNEL variable isn't set within the initramfs.
This is *probably* not an issue.